Move utils in _utils package in aggregation
