+++ src/torchjd/aggregation/_utils/dual_cone.py
