+++ src/torchjd/aggregation/_utils/gramian.py
