+++ src/torchjd/aggregation/_utils/non_differentiable.py
