+++ src/torchjd/aggregation/_utils/str.py
@@ -1,7 +1,7 @@
 from torch import Tensor
 
 
-def _vector_to_str(vector: Tensor) -> str:
+def vector_to_str(vector: Tensor) -> str:
     """
     Transforms a Tensor of the form `tensor([1.23456, 1.0, ...])` into a string of the form
     `1.23, 1., ...`
