+++ src/torchjd/aggregation/aligned_mtl.py
@@ -28,8 +28,12 @@
 import torch
 from torch import Tensor
 
-from ._gramian_utils import compute_gramian
-from ._pref_vector_utils import pref_vector_to_str_suffix, pref_vector_to_weighting
+from torchjd.aggregation._utils.gramian import compute_gramian
+from torchjd.aggregation._utils.pref_vector import (
+    pref_vector_to_str_suffix,
+    pref_vector_to_weighting,
+)
+
 from .bases import _WeightedAggregator, _Weighting
 from .mean import _MeanWeighting
