+++ src/torchjd/aggregation/dualproj.py
@@ -2,10 +2,14 @@ from typing import Literal
 
 from torch import Tensor
 
-from ._dual_cone_utils import project_weights
-from ._gramian_utils import compute_gramian, normalize, regularize
-from ._non_differentiable import raise_non_differentiable_error
-from ._pref_vector_utils import pref_vector_to_str_suffix, pref_vector_to_weighting
+from torchjd.aggregation._utils.dual_cone import project_weights
+from torchjd.aggregation._utils.gramian import compute_gramian, normalize, regularize
+from torchjd.aggregation._utils.non_differentiable import raise_non_differentiable_error
+from torchjd.aggregation._utils.pref_vector import (
+    pref_vector_to_str_suffix,
+    pref_vector_to_weighting,
+)
+
 from .bases import _WeightedAggregator, _Weighting
 from .mean import _MeanWeighting
