+++ src/torchjd/aggregation/mgda.py
@@ -1,7 +1,8 @@
 import torch
 from torch import Tensor
 
-from ._gramian_utils import compute_gramian
+from torchjd.aggregation._utils.gramian import compute_gramian
+
 from .bases import _WeightedAggregator, _Weighting
