+++ tests/unit/aggregation/_utils/test_pref_vector.py
@@ -5,7 +5,7 @@ from pytest import mark, raises
 from torch import Tensor
 from unit._utils import ExceptionContext
 
-from torchjd.aggregation._pref_vector_utils import pref_vector_to_weighting
+from torchjd.aggregation._utils.pref_vector import pref_vector_to_weighting
 from torchjd.aggregation.mean import _MeanWeighting
